Index time synonyms should not remove acronym terms
We had a lot of rules that removed the acronym tokens in favour of
the expanded form. Since we didn't have a corresponding search synonym rule,
this broke searches that used the acronym.

Two ways we can solve this:

1. Add a search time rule to expand the acronym as well (this would artificially increase query length)
2. Contract the full thing at search and index time (wouldn't work for queries that used part of the organisation name)
3. At index time, expand the acronym but keep the original acronym

This commit mostly goes with option 3.

There are some special cases where organisations have been replaced with other ones.
The intent here is that searches for the old thing should include content from the new thing,
but searched for the new thing should just return the new thing.