Ensure that indices are refreshed before checking reindexing output
Add a `refresh` parameter to the reindexing call. This ensures that the
data is refreshed before we compare the data in the new and old indices.

This fixes the integration tests of reindexing, which were failing
intermittently because the new index was not always up to date. It
should also prevent false positives in the comparision check in
production, which we haven't seen but are possible if the index has not
been refreshed.