Remove the my/your synonym
We added this synonym because we thought it would improve queries like
"manage my tax credits", but the healthcheck shows that it has adversely
affected a large number of queries.

We may modify this and add it back in at a later date if we can work out
how to configure it so that it generally improves results.