Make Finder format description indexable
The indexable content field for the Finder format is being [removed](https://github.com/alphagov/whitehall/pull/3575) when publishing occurs. It was felt that the extra keywords added nothing to the ability for these pages to be found on internal search and looking at the most common search terms which end up with visits to those pages confirms this. Therefore we are using the contents of the description field as these Finder's indexable content in Elasticsearch.