Highlight synonym version of title and description fields
The title and description fields should have highlighting applied so
that a user can see which of their search terms appear in the content.
This was not working in the B version of the synonyms A/B test because
the `title.synonym` and `description.synonym` fields did not have
highlighting applied.