Migrate external_content to the govuk index
Map the `external_content` document type to the `recommended-link`
format and migrate it to the govuk index.

The biggest difference between `external_content` and other formats is
that it does not have a `base_path` field because it represents an
external URL rather than a path on GOV.UK.

Instead, the document `link` is generated from the `url` in the details
hash.