Removed shingled query analyzer
This analyzer is supposed to incorporate bigrams (pairs of adjacent words) into
the search query.

This is useful because part of the meaning of a sentence comes from the word
order, for example: "book a driving test for someone else" vs "driving someone
else for a test book".

However, this code never worked as intended because it only analyzed queries.
So a query was broken down into single words and bigrams, but it compared those
tokens to analyzed text that didn't contain any bigrams at all.

This means the bigram part of the query is only functioning as a single word
match. This is very confusing when trying to understand what Rummager is doing.

I've changed it to use the normal query analyzer. This will change results
slightly, because the shingles analyzer didn't include synonyms, but the new
analyzer does.

Bigram matching was implemented properly as part of the 'new weighting' code a
couple of years ago, but it never went live. This is something that could be
revisited in future.