Remove mainstream reference from tests
Replaced with govuk or government depending on what best suited the test.

This attempts to check all tests/code around the migration are working by
changing the code to represent the government migration to govuk. This is to
ensure the code is left in a valid, tested state after we fully remove the
mainstream index.