Work around synonyms that break highlighting
Currently text is tokenized as follows:

- practical driving tests -> {0:practical}    {1:drive}    {2:test}
- practical driving test -> {0:practical,practical}    {1:drive,test}
  {2:test}
- practical test business -> {0:practical,practical}    {1:test,drive}
  {2:busi,test}

This means that:

1. "practical driving tests" isn't being indexed as expected because
it's not an exact match for the rule
2. we are overhighlighting when phrases are replaced by longer queries
at indexing time, because every token in "practical test business"
matches a "practical test" query.

This tries an alternate rule that uses simple contraction, as
recommended here:
https://www.elastic.co/guide/en/elasticsearch/guide/current/multi-word-synonyms.html#_use_simple_contraction_for_phrase_queries

This will stop term positions from overlapping with unrelated tokens.