Fix topic expansion in filters
When search results are aggregated by topic (aka specialist sector),
rummager should expand the topic details to add other fields like title
and content ID. This expanded data comes from a one-off query stored in
a registry.

This expansion broke when we retired the mainstream index because topics
were stored in the mainstream index and the registries code had _not_
been migrated to use the govuk index.

This broke pages which rely on topic aggregation, such as
https://www.gov.uk/government/organisations/environment-agency/services-information

All other registries use the government index (which has not yet been
migrated to govuk), so they can be left as they are.