Add expanded_search_result_fields to residential_property_tribunal_decision schema
This allows rummager to translate the computer-readable values from the
document into human-readable values when it presents the search results as
json.  This may be a bit of tech debt that we should clean up as we
thought this config was deprecated.  The approach under discussion is that
we either:

1. Push both human and computer readable values into the documents
   provided to rummager
2. Make finder-frontend do the same as government-frontend and use the
   facet definitions of the finder to do the translations for us.