Zip page traffic data on sidekiq
As we need to process the entire index onto the sidekiq queue we have added
zipping which see a reduction of about 60% in the amount of data put into
memory. This solution has been coded into teh actual worker so it is easy to
see the deflate/inflate methods being called.