Revert using foreman in dev environment
This removes foreman from the Gemfile and instead installs foreman
separately via `gem install` in the Dockerfile so that it can be used in the
docker environment.

The reason for this is that running everything via unicorn in dev can
cause some wtfs (such as breaking better_errors) and installing foreman
via Gemfile does contradict some contenious advice given by the foreman
gem maintainer.