Remove Sentry logging of InvalidFormatError
The invalid format error is an extremely common error - it is logged in
Sentry 152k times and we (those working on Publishing E2E Tests) see it
all the time in our logs.

Since this is not actually an exceptional circumstance and rather the
norm until something is set up then this shouldn't be logged in this
manner as it's unlikely to be providing any value.