Add mutual synonym of "Asad" and "Assad"
Departments and agencies other than FCO tend to publish content
referring to the President of Syria with the spelling “Assad”. FCO use
the agreed form “Asad”. Users need to see both sets of content when
searching for either spelling.

This is a synonym for both searching and indexing so that combined
frequencies of the terms are taken into account, which will make search
perform better.