Make news_story indexable when published by Content Publisher
The news_story document type will soon be publishable by Content Publisher as
well as Whitehall. Whitehall indexes its content to Rummager directly, whereas
Content Publisher will use the conventional way of relying on Publishing API to
take care of communicating with Rummager. In order to make it possible for
Content Publisher content to be indexable, we whitelist news_story as an
indexable format and specify Content Publisher as the publishing app for it.
news_story remains in the non_indexable list as well and we specify Whitehall
as the publishing app here. This approach will allow us to check the publishing
app for a piece of content and use this to determine whether it should be
indexed or not.

Note we had to amend an existing test in migrated_formats_spec to convert
the returned hash from `news_story` into an array in order for the test to work
correctly.