Allow filtering with topical_events
This makes it possible to search for documents related to topical_events.
For example: rummager.search(filter_topical_events: 'budget-2018')

topical_events are currently slugs like budget-2018, not the full topical event resource.

This change enables some Whitehall finders to use rummager.

Related Trello card: https://trello.com/c/FSMKn70m/26-make-it-possible-to-filter-rummager-by-topical-events.

This currently is an incomplete task, as we will need to populate the topical_events field
we will need to also make a change in Whitehall and possibly also the publishing-api in
order to make filtering by topical events useful.