Cap the limit of the "start" param
Values that are greater than around 900,000 result in a 500 error.  We
currently have around 375,000 results in Rummager, so capping the maximum
start value at 500,000 doesn't seem too bad.

We've experiencing repeated scanning requests to search pages which result
in numbers (eg 29999800) that are vexatious.  Platform Health have a card to
relieve this in the longer term, but this will hopefully reduce the number
of 500s we serve in the meantime and reduce the noise on 2nd line.

I've removed the test for expecting a 400 when submitting large integers to
elasticsearch because in practice this results in a 503.  We're also putting
defences in front of that now.