Introduce any/all filtering of multivalued fields.
Add filter_all_<field>, filter_any_<field>, reject_all_<field> and
reject_any_<field> to query multivalued fields.

When filtering multivalued data, previously the behaviour was to
select documents that have at least one of the specified values.

This PR adds an _all_ operation that allows to filter documents that
contain all of the specified values. The _any_ operation is the
previous behaviour and is the default.