Index documents tagged to world locations in govuk index
Previously, all tagging to world_locations were handled in whitehall.
When a document is published whitehall would update the government index
in rummager which is the older index. The search endpoint would return
results from both the new (govuk index) and the old (government index).

Content-publisher can now tag world_locations to a document however the
index it populates (govuk index) does not index this field currently,
this change enables documents published by content publisher which are
tagged to world_locations to be indexed so the search endpoint can
return these results.