Expand World Locations in Finder responses This fetches world location data from https://www.gov.uk/api/world-locations and uses it to provide titles and slugs to Finders for use in faceting with world locations.
We need to use an external source as most world locations were removed from
rummager in this commit: 45ce7295a731991f483f5ba238c3424549213746.
https://github.com/alphagov/whitehall/pull/3966/commits/45ce7295a731991f483f5ba238c3424549213746

This is a short term fix. The longer term solution would be to bring world locations back
into the content-store and rummager. That may be tricky and is being discussed at the moment.

Only the title and slug of a world locations is required for Finder-frontend.

This is part of the migration of finders from whitehall to finder-frontend.
After this is merged we can add world_locations as an available filter to
finders in finder-frontend.