Amend tagging task to process csv
We have decided that content to be tagged will be provided in a csv.
The format will be:

```
base_path, all, sector_business_area, employ_eu_citizens,....
```

Individual items in the facet columns will be comma delimitted facet
values in a sluggified format.

When `all` is populated with "yes", we want to tag to every facet, every
value.