Check image_url in Elasticsearch Presenter
This checks if an image is available in the payload and, if not, looks up the
document's organisation's default news image and uses that in the payload
instead. This approach helps reduce instances where a frontend app uses a
default image (which is vulnerable to breaking) in the absence of an available
image.