Expose image_url for any document type
Since https://github.com/alphagov/rummager/pull/1354 lots more things are able
to have an image.

We need to make this change so that rummager can expose the information. For example
https://www.gov.uk/api/search.json?filter_link=/guidance/the-highway-code now has
an image_url.  When this document is retrieved and includes the `image_url` field
an [error is raised](https://sentry.io/govuk/app-rummager/issues/790236958/?query=is:unresolved)
because the schema that defines the-highway-code is not `edition.json`.

By adding the field to the base_elasticsearch type, this error should be
resolved.