Add EU Exit finders for individuals
Adds finders for each of the top level taxons. We're doing this to simplify
the code we need to write in finder-frontend to build queries.

We haven't included the Government taxon because the query doesn't seem to
work because the citizen taxon is a child of Government.

We don't want to include email signups yet as they're not
ready.

This has a base filter of the citizen topic under /government/brexit
which is draft at present, and the relevant topic.