Initial warden configuration for GDS-SSO
This installs the necessary gems (warden, oauth2 and warden-oauth2) to
communicate with signon for SSO.

It does not use the gds-sso gem as this has expectations of being in a
Rails application with an ActiveRecord-like database storage layer,
whereas Rummager is a Sinatra application without a database persistance
storage layer.

We're anticipating that we don't actually need to store the users that
are authenticated in Rummager so a simpler process can be used.

Paired with Karl Baker