Add cache service to Rummager
This sets up using ActiveSupport::Cache for use with SSO. This is so we
can store a bearer token lookup for 5 minutes. Without this signon is
sent many requests.

We use a null cache store in non production environments to not cache
anything, whereas in production we use a memory store. Normally Rails
uses a FileStore cache as this persists between starts of the
application, however it seems simpler to use a memory store here and
only to change that if the cache needs change.