Set up warden strategies in rummager app
This sets up a mock strategy for authentication which respects the same
environment variables as the GDS SSO gem.

It moves all the file requires into lib/rummager.rb

It sets up authentication inside the sinatra app which resolves an issue
that we need access to the rack `use` method (this previously worked for
running the app but not for running the tests)

and finally it decodes the signon json response to return a hash of a
user with a similar one in the mock variant.