Remove document type from prepare brexit finder
This doesn't appear in the search results as the citizen topic isn't live
(currently `alpha`). It's misleading to include this in the query, so we
should remove it.  There's a link to the foreign travel advice page on the
[going and being abroad finder](https://www.gov.uk/prepare-eu-exit/going-and-being-abroad)
anyway.

If we want non-whitehall content to appear in the finders, we need to set the
topic to `live` in [content tagger](https://content-tagger.publishing.service.gov.uk/taxons/d7bdaee2-8ea5-460e-b00d-6e9382eb6b61)

There's been some resistance to doing this at present because it makes it
available for publishers to tag to in Whitehall.