Bump the total mapping fields limit to 2000
Elasticsearch 5 imposes a 1000-field limit for schemas.  So does
elasticsearch 6.  However, that limit is just default configuration,
it's not a hard limit.

The point of this limit is to avoid an explosion of fields (and so
memory usage), particularly with dynamic indices:

> The following settings allow you to limit the number of field
> mappings that can be created manually or dynamically, in order to
> prevent bad documents from causing a mapping explosion

We have more than 1000 fields.  This isn't a problem as we have a
relatively unchanging set of schemas, so the number of fields isn't
going to suddenly blow up any time soon.  Let's just increase the
limit and have one less thing to worry about.

I tested this by bumping the limit to 2000, and seeing that it had no
significant effect on resident memory usage (as expected: it's the
fields that use memory, not the limit).