Tag metadata on change notification messages
Change notification messages for links updates were causing
the DocumentPreparer to return a version of the document without
additional metadata from the Indexer::MetadataTagger. This meant
that the import process and PublishingEventWorker updates would
correctly assign the metadata but a change notification from link
updates would rewrite the document back to the default unaugmented
state. Arguably this means we should make the default state consistent
with the results of tagging with metadata but we've yet to decide on
the correct strategy for tagging the documents in question. So
for now fix this code-path to be consistent with PublishingEventWorker.