Add Content Item Publisher module
This new module handles the presentation and publication
of raw YAML / JSON files in rummager into content items
sent to the publishing api.

It is intended to be called by a rake task (to be added).

The intent is to supersede lib/publishing_api_finder_publisher.rb.
This is for two reasons:
a) email signup content items are growing sufficiently complex
that they deserve their own classes for publication / presentation.
b) We dont want to remove publishing_api_finder_publisher.rb yet;
this is in use by other teams and other finders depend on this.

The plan is to add this module and an additional rake task
for publishing finders and their email signup pages while
keeping the old method alive.