Use a different content ID for taxon removal test
This same content ID is used in the external content tests.  We use
the content ID as the document ID in these integration tests (normally
we use the base path).  As we're now using a single mapping type, this
means the document identifiers collide.

We use versioning to ensure that a document in elasticsearch hasn't
changed while we attempt to delete it.  I assume inserting a second
document with the same ID is altering the version number, and so
causing the delete to fail.

I don't think this problem can arise during normal execution, as it
requires two documents to have the same ID, and the publishing-api
prevents that by ensuring that both base paths and content IDs are
unique.