Add business finder email config
The default email signup created when the finder is published creates a
list of `email_filter_facets`. We need to replace this with
`email_filter_by`, `email_filter_name` and `email_signup_choice` so
that users are not forced to select a facet when they sign up for email
alerts.