Have Rummager wait while updating the popularity
Currently updating the popularity values involves putting the
documents in the index on to the Sidekiq queue, so that they can be
sent back to Elasticsearch with the updated popularity value.

This uses lots of memory, and recently, enough memory to cause Redis
to run out.

I've attempted to switch this to use the update, rather than index
action, but this doesn't work due to the use of `external_gte`
versioning, as the update increments the version, which could cause a
problem with processing messages from the Publishing API. ADR 002
describes one way to workaround this issue "External versioning with
multiplier", but changing the versioning scheme seems like a big
change to me.

Instead, this commit changes the popularity updating Rake task to
block while the jobs are being processed. As the nightly search
analytics job updated 3 indexes, I'm hoping that this will reduce the
peak Redis memory usage.