Add new rake task to publish business finder
`publish_document_finder`, the original task used to publish the
business finder, auto-generates and publishes the email signup content
item, by using attributes from the finder content item.

That means that for the business readiness finder, users are being
forced to select facets when they sign up for email alerts.
We need to be able to specify and publish different config for the email
signup page to allow users to subscribe to all content in the finder.

There is an existing task that does something similar, `publish_finder`,
however that task assumes all finder config is stored in
`config/finders`, whereas the business readiness finder config is copied
into `config/` by the deployment scripts.

We could:
1. Update `publish_finder` to take a “FINDER_PATH” argument, but this
  makes the task more difficult to run for users.
2. Change the deployment scripts to copy business readiness finder
  config in to `config/finders`, but the deployment scripts also copy
  the config to email-alert-api, so that code would have to be updated
  too.
3. Add a new rake task that does the same thing as `publish_finder`, but
  hard-codes the location of the business readiness config. This
  duplicates some of the code, but has the advantage of being easy to
  remove when the time comes, and easier for users to run.