Expand all synonyms
There are two types of synonym definitions in elasticsearch.

Explicit mappings:

    a, b, c => d, e, f

Which means "match any token sequence on the LHS and replace with all
alternatives on the RHS".

Equivalent synonyms:

    a, b, c

Whose behaviour depends on the `expand` parameter in the schema.  If
`expand` is true (the default), it is equivalent to:

    a, b, c => a, b, c

If `expand` is false, it is equivalent to:

    a, b, c => a

We had trouble with equivalent synonyms not being expanded.
Explicitly setting `expand` to true in the schema didn't change
anything, but writing the equivalent synonyms in the explicit
`expand`-is-true form did.  Not really sure what's up with that.