Allow AND query across facet values
The problem with moving to the new `facet_values`
type of query, instead of the existing metadata facets,
is that Rummager treats queries on the same column
as `OR` queries.

All of the existing facets are in different fields,
the `facet_values` approach makes this a single field.

This commit introduces a second copy of the `facet_values`
called `and_facet_values`

When populated with identical data this will allow queries like:

&filter_facet_values=uuid-from-first-facet&filter_and_facet_values=uuid-from-different-facet