Add healthcheck endpoint
This adds a healthcheck endpoint, which returns a response
that can be used by Icinga to create alerts. At the moment
we have two alerts:
- Sidekiq cannot connect to Redis
- Sidekiq queue latency is beyond a certain threshold

The first alert is a new one, which will help us if the redis
queues arent accessible.

The second alert is designed to replace the Icinga alert in
govuk-puppet which looks at graphite to create the alert.
I will remove that alert once this one is up and working.

Once this is merged, I will make a change to govuk-puppet
to make Icinga poll the search-api /healthcheck endpoint
and create alerts using the result.

Trello: https://trello.com/c/cNvfdZfU/372