Allow queries to contain nested conditions
Updates user filters to allow a parameter structure
which can denote and/or logic in the form of nested
values.
eg. `{"and"=>{"0"=>["a", "b"], "1"=>["c", "d"]}}`
The keys within the `and` hash are disregarded, these
are necessary to construct a Rack compliant set of request
parameters which can be accurately separated into and/or
clauses.