Adds task to auto generate finder signup facets
There are quite a few facets in the business finder,
making the facets for the signup content item would
be laborious and error prone.
Creating a rake task ensures that it is correct and that
there is a single source of truth.

Run it with something like:
rake convert_finder_facets_to_signup_facets[/find-eu-exit-guidance-business,signup.yml]