Add metric for query build time
Building a query involves two round-trips to elasticsearch: one to get
best bets, another to look up the real name of the "govuk" index.
Recording query build times is helpful because that + the raw search
time is (in principle) the fastest search-api could possibly be, if
all other overhead were eliminated.