Count search requests over the standard endpoint
We count the number of searches using batch searches (currently nil!)
but it's nice to be able to see the number of searches using the standard
endpoint too.

It's allowed us to identify slow running machines in the past.

(For example in this dashboard https://grafana.production.govuk.digital/dashboard/db/search-requests-and-errors?panelId=3&fullscreen&orgId=1)