Load page traffic into all active clusters
This loads page traffic data into all indexes,
not just the primary one.

There's possibly better ways to do this (e.g.
copying from one index to another?). But this
is consistent with the way data enters into all clusters.

One thing that might be nice is to create a worker
for each cluster.