Add SearchConfig.default_instance
This allows us to add more parameters in the future, and code which
uses the default won't need to add all the extra arguments.  It also
communicates the intent better than explicitly using
Clusters.default_cluster.

The reason this is a new method rather than making the clusters
parameter of SearchConfig.instance optional is so that you have to
explicitly opt in to using the default configuration, to reduce
mistakes.