Adds topical event as a filterable parameter
Topical events have links to publications and announcements which redirect
to these finders.  Until this change, the parameter would just be ignored.

This does not show a facet or a facet tag for topical events, but purely fixes
the issue in https://govuk.zendesk.com/agent/tickets/3732860.

We'll address this more fully with planned work.