Set a default ELASTICSEARCH_B_URI when testing
We have two elasticsearches in the dev VM, in docker, and on CI.
Given that there are tests which touch the cluster logic, we want to
make use of all of them!

An alternative would be setting the environment variable in CI.  I
decided against that because that the tests are currently failing when
run against multiple clusters, so adding the variable in CI would
break builds of every branch.