Add global caching object
Currently most caching of large configuration files is done by using
memoization in several classes.

This commit adds a very simple caching solution so that all memoizations
can be done in this one class.

This has as an advantage that caching can be controlled centrally
and that before each test we only neeed to clear this one cache
to ensure that all caching throughout the app has been cleared