Remove memoized caching in objects and move some to the Cache class
This commit removes unnecessary memoization in classes because they
can cause issues in regards to isolating unit tests.

Where caching is necessary, it is moved to the one central caching
class which is cleared before each test starts

A number of tests have been modified so that they do no longer depend
on memoization.