Remove brexit, eu exit synonym
We've decided that this is no longer helpful for users.  Existing
guidance is being updated to use "Brexit" instead of "EU Exit", and
cases where "EU Exit" remains are places like official names of
publications (eg, "The Customs (Export) (EU Exit) Regulations"), in
which case it's being used in a specialist sense.