Add stats announcements display date
This is currently only available as a property of the [metadata field](https://www.gov.uk/api/search.json?filter_format=statistics_announcement&fields=metadata),
which makes it tricky to expose using the "finder" pattern which is optimised
for retrieving properties at the top level of a result.

Examples of current values for this are:
- "September to October 2019"
- "30 January 2020 9:30am"
- "October 2019"

I've named it fairly generically so that this field may be reused by things
other than statistical announcements in future.  This is consistent with the
`release_timestamp` field above it in the definition file.