Use match_all_terms in unquoted query, not match_phrase
match_phrase is inappropriate for unquoted queries where we're not
doing exact matching, because we don't have a slop parameter set.
This means that terms can't be reordered or have other terms between
them, so the match_phrase will actually match only very rarely.

For example, "brexit business" won't match_phrase the title of the
business readiness finder, which is "Find Brexit guidance for your
business", because there are words between "brexit" and "business".

It's possible that we do want to rank phrase matching higher, in which
case we may want those clauses here, but I suspect in the general case
match_all_terms will be far more successful.