Enable automatic slicing on Elasticsearch reindexing
This PR enabled slicing on the reindexing migration task, which in
localised testing provides a 20% performance improvement. This
improvement scales based on the number of shards available to
an index, so should be higher still on our Elasticsearch cluster.

Slicing enables Elasticsearch to split large scrolls into multiple
chunks, processing each one in parallel. It is safe and relatively
memory efficient, and should lead to no detriment to performance
even in cases where we see no gain.