Turn off RSpec cops
Previously we enabled RSpec cops, but have since disabled several of
them globally, as well as many times in isolated parts of specs. This
disables the RSpec cops again to avoid having to do a large number of
manual corrections as part of turning off diff linting. It also seems
reasonable to do this if they weren't having a consistent impact on the
code we've been writing.