Add popularity c variant
This adds view_count to content indexes

The view count is already in the popularity index, but not
specified as an indexed field. This makes it available in all
indexes for query time popularity boosting.

This uses view_count for query time boosting, to
see if using the absolute view count rather than
popularity rank will have an impact on relevancy.

Trello: https://trello.com/c/pNxIeTFc/1044