Highlight quoted query fragments in titles and descriptions
Highlighting is done on the title and description fields (or their
.synonym fields), but quoted queries are run against the .no_stop
fields.

One approach would be to add the .no_stop fields to the highlighting
query, and then using that field in Search::HighlightedField if the
other fields aren't present.  But that wouldn't help with queries that
mix quoted and unquoted fragments.

For queries which match quoted and unquoted fragments, we want to
combine highlighting from two different fields.  We can do this by
using a (much simpler!) query to highlight results.  I've used
"match_phrase" for quoted fragments (so they're highlighted in their
entirety) and "match" for any unquoted fragment (so individual
keywords are highlighted).

Highlighting quoted fragments isn't perfect:

- Stopwords aren't highlighted
- Synonyms are applied (so a quoted query could match something new)

I'm not sure how best to work around those issues.  We could do all
highlighting against the no_stop field and apply all the index-time
synonyms at query-time for the unquoted fragment.  But this way is
better than not working at all.