Use HTTParty to make rank eval request
We need to make a basic http request here rather than
using the elasticsearch ruby client.

This is due to two factors:
- AWS Elasticsearch service has a bug that prevents
calling the rank_eval API without an index. We've raised
a ticket with them but have not been given an ETA on when
this will be fixed.
- The Elasticsearch ruby client has a bug that prevents
specifying an index when using the API. We raised a PR
to fix this (https://github.com/elastic/elasticsearch-ruby/pull/724).

This makes the call specify all indexes '*' and doesn't use the
elasticsearch client library. Once either issue is fixed
we can remove this.

Trello: https://trello.com/c/Uq13lWgG