Learning To Rank
Trello: https://trello.com/c/P3pynH7k/1092-spike-validate-learn-to-rank

This spike demonstrates how we can implement [Learning To Rank](https://en.wikipedia.org/wiki/Learning_to_rank)
using Tensorflow’s [ranking module](https://github.com/tensorflow/ranking).

The spike is functional, but not ready for merge.

Contents:

1. What is Learning To Rank?
2. Implementation
3. Results
4. Next steps

What is Learning To Rank?

Learning To Rank (LTR) for our use case is the application of machine learning
to create a ranking model that can be used to improve search relevance.

This is a common approach for advanced search applications.

There is a good explanation on what LTR is here:

- https://elasticsearch-learning-to-rank.readthedocs.io/en/latest/core-concepts.html
- https://opensourceconnections.com/blog/2017/02/24/what-is-learning-to-rank/

It is easiest to explain with an example:

Say you searched for 'Harry Potter'. Search API might return
the following results:

1. HMRC Sign in
2. Harry Potter World
3. Pottery barn
4. Harry Potter

These are not good results.

How do we improve search results currently?

Our approach now would be to manually rank these results,
giving them a score of between 0-3 (0 being irrelevant, 3 being
completely relevant):

```
query         rating    document
harry potter  0         HMRC Sign in
harry potter  2         Harry Potter World
harry potter  0         Pottery barn
harry potter  3         Harry Potter
```

These judgements are an expression of the ideal ordering.

We can then use nDCG to tell us how good the results are.
We might get a score of 0.3 for this result (from 0 to 1).

We could then investigate why the results are bad. Perhaps
there is a best bet. Or popularity might be pushing HMRC
docs. We can fix these things, then re-run nDCG, and
we might get 0.7 for this query. Good stuff.

This works, but doesn't scale well. Eventually this process
becomes a game of whack-a-mole. If we downgrade popularity
boosting, then this query might have better results, but
other queries might have worse results.

Hand-tuning relevance doesn't work very well because
field boosts are complex, and difficult to maintain.

Learning to Rank makes tuning relevance:

- automatable
- scalable
- more successful
- more maintainable (arguable)

LTR won't make relevance tuning easier. But it could make search
results more relevant.

How do we improve search results with Learn To Rank?

The end goal is to make search results more relevant, without
manually increasing the boost of a field.

We would go through the same process of rating the search results:

```
harry potter  0         Pottery barn
harry potter  3         Harry Potter
...
```

However, after doing this, we would train a model with this data
to rank the results better. The model would optimise for nDCG.

We would provide the model features, such as the fields we boost on.

For example, take the SVM dataset below for the query 'harry potter' (qid:1)

```
0 qid:1 popularity:0.3 title_score:0.2 description_score:0.3 recency:0.9 pagerank:5 ...
3 qid:1 popularity:0.1 title_score:0.6 description_score:0.7 recency:0.5 pagerank:2 ...
```

This data, in addition to many other graded query:document pairs
and their features can be used to train a model.

The model might learn that title and description scores are
slightly more useful that popularity and recency, and how much,
in certain circumstances.

Implementation

I chose to use [TensorFlow's ranking library](https://github.com/tensorflow/ranking)
for this.

TensorFlow is well supported, and the ranking library is under active
development.

I wanted to try https://github.com/o19s/elasticsearch-learning-to-rank, but
as we are using AWS Elasticsearch, we are not permitted to install custom plugins.

Architecture:

User searches
Search API pulls results out of Elasticsearch
Search API send results to LTR model served by TF Serving
LTR model returns results re-ranked
Search API returns results to user

The model is trained on implicit and explicit relevancy judgements. Ratings
are provided by manual raters and derived from user click data.

Features can added and improved on.

Notes on the code in this spike:

Within this spike there is a new module `LearnToRank` which contains
various utilities for training and using the ranking model.

We needed to rewrite NDCG outside of Elasticsearch because the re-ranking
happens outside of ES.

`LearnToRank::Reranker` is called after the results have been extracted from
Elasticsearch, it calls the separate process serving the model which
re-ranks the results.

There is also a directory of scripts for training and serving the model.

Results

This spike contains a working MVP of a ranking model that can be
trained with relevancy judgements to re-rank results at query-time.

The results are positive. For our existing dataset (explicit relevancy
judgements) the model gets a higher score (9.2) compared to the original (0.89).
This was done on local data, we might see different results in e.g. integration.

For the combined dataset (explicit and implicit sourced from GA) we
get a score of 0.88 with the model and the same score without. The
model has a very slightly higher score (0.88265 vs 0.88011).

I think this demonstrates that with some more work (data/feature
engineering), the model can outperform our existing search algorithm.

The model uses ndcg as one of the measures by which it trains itself,
so it's not surprising that it can get a better score.

We were able to beat our existing search's ndcg score within a
few days. This doesn't mean that search is better, just better
by this metric with these relevancy judgements available.

In Bloomberg's talk on their project, they devote half of the talk
to discussion of how they made query performance better after implementing
LTR. I was worried that our search performance might also be hit.

The re-ranking model is very quick. TF Serving is production ready;
I think any performance bottlenecks will be in ruby and Elasticsearch.

I didn't notice a performance hit, but I haven't done proper testing of this.

The code is not much more complex. Architecturally the application is more
complex. It requires a separate process to be running. It requires some
knowledge of machine learning concepts and Learn To Rank from developers
and product managers.

It's harder to do, but will lead to better results. I am unclear whether
this lead to a harder to maintain application.

Pros:

- Potentially significant improvements to relevancy
- This is a well trodden path, as many orgs use LTR (Airbnb, Wikimedia,
  commercial search engines, Bloomberg, Yelp, etc.)

Cons:

- Added architectural complexity
- Harder to build and debug in contrast to hand-tuning

Next Steps:

Aside from making this more robust so that it is production ready,
I see two main areas for further development:

- Further research into how we should implement Learn To Rank
- Feature engineering
- Acquiring more/better relevancy judgements

The tasks would be aimed at making the model give more precise
relevancy predictions, should we decide to proceed with LTR.